first version of these files.
this producers uses the method provided by Alex and Renaud to
make list of uncalibrated rechits.

the code also accesses the offline db to read pedestals and weight matrices.
in this version the code to convert vector< vector<EcalWeight> >
into CLHEP matrices is implemented as a private method. The code
should be eventually moved to a utility class.